Fixing one more failed copy
